requirements.txt
# 智库工作流自动化系统依赖包
# 本小姐精心挑选的依赖哦！(￣ω￣)ノ

# AI 相关
openai>=1.0.0              # OpenAI API
anthropic>=0.18.0          # Anthropic Claude API
requests>=2.31.0           # HTTP 请求（用于本地模型）

# 配置文件处理
pyyaml>=6.0.1              # YAML 配置文件解析

# 日志和工具
python-dateutil>=2.8.2     # 日期处理
pathlib>=1.0.1             # 路径处理（Python 3.4+ 内置）

# 可选依赖（用于高级功能）
# jieba>=0.42.1            # 中文分词（用于更好的关键词提取）
# scikit-learn>=1.3.0      # 机器学习（用于相似度计算）
# numpy>=1.24.0            # 数值计算
